order schedule

+++ server/controllers/transectionCtrl.js
@@ -42,8 +42,13 @@ const deleteTransection = async(req,res) => {
   //check authentication
   if (req.current_user != null) {
     try {
-      await transectionModel.findOneAndDelete({_id:req.body.transacationId});
-      res.status(200).send('Transaction Deleted');    
+      if(req.body.isAuto ==1){
+        res.status(400).send("Can't delete Auto inserted Data"); 
+      }else{
+        await transectionModel.findOneAndDelete({_id:req.body.transacationId});
+        res.status(200).send('Transaction Deleted'); 
+      }
+         
     } catch (error) {
       console.log(error);
       res.status(500).json(error);
@@ -60,9 +65,14 @@ const editTransection = async(req,res) => {
   //check authentication
   if (req.current_user != null) {
     try {
-      // writing query in mongo db 
-      await transectionModel.findOneAndUpdate({_id:req.body.transacationId},req.body.payload);
-      res.status(200).send('Edit Successfully');
+      if(req.body.isAuto ==1){
+        res.status(400).send("Can't update Auto inserted Data"); 
+      }else{
+        // writing query in mongo db 
+        await transectionModel.findOneAndUpdate({_id:req.body.transacationId},req.body.payload);
+        res.status(200).send('Edit Successfully');
+      }
+      
     } catch (error) {
       console.log(error);
       res.status(500).json(error);
@@ -81,7 +91,8 @@ const addTransection = async (req, res) => {
     try {
       let data = req.body;
       data.userid = req.current_user.user_id;
-      const newTransection = new transectionModel(data);
+      data.isAuto = 0;
+      const newTransection =
       await newTransection.save();
       res.status(201).send("Transection Created");
     } catch (error) {
